// src/types.ts
export interface SatelliteImageData {
  name: string;
  url: string;
  width: number;
  height: number;
  sizeFormatted: string;
  file?: File;
  loadedAt: string;
}